improve batch function

+++ NeuralNetwork.cpp
@@ -14,12 +14,10 @@ NeuralNetwork::NeuralNetwork(const bool load_weights) {
 }
 
 void NeuralNetwork::doOneBatch(const int current_batch_index, const std::vector<DataPoint> &dataset_train) {
-
-    const int start_index = current_batch_index * BATCH_SIZE;
-    const int end_index = std::min(((current_batch_index + 1) * BATCH_SIZE), static_cast<int>(dataset_train.size()));
+    const size_t start_index = static_cast<size_t>(current_batch_index) * BATCH_SIZE;
+    const size_t end_index = std::min(static_cast<size_t>(current_batch_index + 1) * BATCH_SIZE, dataset_train.size());
 
     const WEIGHT_SHAPE delta_matrix = getDeltaMatrix(start_index, end_index, dataset_train);
-
     adjustWeights(delta_matrix);
 }
